+++ kedro_vertexai/hooks.py
@@ -1,8 +1,5 @@
 import os
-import warnings
-from typing import Iterable
 
-from kedro.config import ConfigLoader
 from kedro.framework.hooks import hook_impl
 
 from kedro_vertexai.constants import (
@@ -11,8 +8,6 @@ from kedro_vertexai.constants import (
     VERTEXAI_JOB_NAME_TAG,
     VERTEXAI_RUN_ID_TAG,
 )
-from kedro_vertexai.context_helper import EnvTemplatedConfigLoader
-from kedro_vertexai.runtime_config import CONFIG_HOOK_DISABLED
 from kedro_vertexai.utils import is_mlflow_enabled
 
 
@@ -31,23 +26,4 @@ class MlflowTagsHook:
                 mlflow.set_tag(VERTEXAI_JOB_NAME_TAG, job_name)
 
 
-if not CONFIG_HOOK_DISABLED:
-
-    class KedoVertexAIConfigLoaderHook:
-        @hook_impl
-        def register_config_loader(self, conf_paths: Iterable[str]) -> ConfigLoader:
-            return EnvTemplatedConfigLoader(conf_paths)
-
-else:
-
-    class KedoVertexAIConfigLoaderHook:
-        pass
-
-    warnings.warn(
-        "KEDRO_VERTEXAI_DISABLE_CONFIG_HOOK environment variable is set "
-        "and EnvTemplatedConfigLoader will not be used which means formatted "
-        "config values like ${run_id} will not be substituted at runtime"
-    )
-
 mlflow_tags_hook = MlflowTagsHook()
-env_templated_config_loader_hook = KedoVertexAIConfigLoaderHook()