+++ tests/test_context_helper.py
@@ -45,14 +45,70 @@ class TestContextHelper(unittest.TestCase):
         )
         metadata = Mock()
         metadata.package_name = "test_package"
-        context = MagicMock()
-        context.config_loader.return_value.get.return_value = ["one", "two"]
-        with patch.object(KedroSession, "create", context), patch(
-            "kedro_vertexai.context_helper.EnvTemplatedConfigLoader"
-        ) as config_loader:
-            config_loader.return_value.get.return_value = cfg.dict()
+        session = MagicMock()
+        session.load_context().config_loader.get.return_value = cfg.dict()
+        with patch.object(KedroSession, "create", return_value=session):
             helper = ContextHelper.init(metadata, "test")
             assert helper.config == cfg
+
+    def test_config_with_omegaconf(self):
+        from kedro.config import OmegaConfigLoader
+
+        with TemporaryDirectory() as tmp_dir_raw:
+            tmp_dir = Path(tmp_dir_raw)
+            (tmp_dir / "conf" / "base").mkdir(parents=True, exist_ok=False)
+            (conf_dir := tmp_dir / "conf" / "local").mkdir(parents=True, exist_ok=False)
+            cfg = PluginConfig(
+                project_id="test-project",
+                region="test-region",
+                run_config=RunConfig(
+                    image="test-image", experiment_name="test-experiment"
+                ),
+            )
+            (conf_dir / "vertexai.yml").write_text(yaml.dump(cfg.dict()))
+
+            metadata = Mock()
+            metadata.package_name = "test_package"
+            session = MagicMock()
+            session.load_context().config_loader = OmegaConfigLoader(
+                str(tmp_dir / "conf"),
+                config_patterns={"vertexai": ["vertexai*"]},
+                default_run_env="local",
+            )
+            with patch.object(KedroSession, "create", return_value=session):
+                helper = ContextHelper.init(metadata, "test")
+                assert helper.config == cfg
+
+    @unittest.expectedFailure
+    def test_config_empty(self):
+        metadata = Mock()
+        metadata.package_name = "test_package"
+        session = MagicMock()
+        session.load_context().config_loader.get.return_value = None
+        with patch.object(KedroSession, "create", return_value=session):
+            helper = ContextHelper.init(metadata, "test")
+            _ = helper.config
+
+    @unittest.expectedFailure
+    def test_config_raises(self):
+        metadata = Mock()
+        metadata.package_name = "test_package"
+        session = MagicMock()
+        session.load_context().config_loader.get.side_effect = ValueError()
+        with patch.object(KedroSession, "create", return_value=session):
+            helper = ContextHelper.init(metadata, "test")
+            _ = helper.config
+
+    @unittest.expectedFailure
+    def test_config_invalid(self):
+        metadata = Mock()
+        metadata.package_name = "test_package"
+        session = MagicMock()
+        session.load_context().config_loader.get.return_value = None
+        session.load_context().config_loader.__getitem__.side_effect = KeyError()
+        with patch.object(KedroSession, "create", return_value=session):
+            helper = ContextHelper.init(metadata, "test")
+            _ = helper.config
 
 
 class TestEnvTemplatedConfigLoader(unittest.TestCase):